CMR-4067: Replaces use of Throwable to quit queue reader worker.
Replaces us of Throwable to quit queue worker after tests with
ExceptionInfo created with call to `ex-info`. Added to provide
a more distinct 'quite message' after test run. Also updated
docstring to provide an explanation of the use exception handling
to keep the worker alive and to kill it after test runs.