CMR-8614: Remove the mime-type from the SQL database, have Generic code treat mime-types in a standard way (#1733)
* CMR-8614: database migration to remove mime_type column

Adds a migration to remove or restore the mimetype column in the oracle database. Removes references to mime_type in the codebase; removes the test condition that requires mime_type

* CMR-8614: Removing mime-type from the grid schema

Removing the mime-type from the grid schema. The other generic schema do not have mimetype as a field

* CMR-8614: fix comment, remove unused module requires

* CMR-8614: returning mimetype to the schema, reversing 086

In 086 db migration the up function and the down function should have their definitions switched. The mime_type should NOT be removed from the grid schema