Cmr-8665: Adding dynamic generic docs (#1731)
* CMR-8665: Dynamically insert generic docs

The generic docs are dynamically added to the ingest and serach api's based on the markdown files that are present in the generic schema's directory i.e. schemas/<generic-concept-name>/<version>/<markdown to be inserted>

* CMR-8665: html template for table of contents

Moved html templates to /resources dir and they are being slurped instead. Added tests, altered grid markdown. Resolving dependency issues with the inflection module

* CMR-8665: cleaning up

Removing unused code improving documentation; fixing alphabetical order for project import

* CMR-8665: Removing doc type context from common-lib

Using the page we can dynamically pass the name of the document from which we want to retrieve data for depending on which app (search or ingest)  is calling the docs-routes function

* CMR-8665: ingest table of contents working

* CMR-8665: Changing markdown files to add in generic toc

The API.md files need to have markdown added to them so that the format is fixed

* CMR-8665: Table of contents and body working

Table of contents and body are working; writing tests, and cleaning up is next

* CMR-8665 writing tests cleaning up

Writing some tests; cleaning up debugging statements

* CMR-8665: Adding more information for functions; cleaning up

Remove the TOC docs they should not be included anyway as things are being dynamically generated.

* CMR-8655: fixing some some spacing;formatting

Using the existing read-generic-doc-file to slurp the .md file

* CMR-8665: adding another debug print; removing a search.md

Search.md should be removed for serviceentry it was for local testing and should not be committed. reverting reader call

* CMR-8665: polishing; removing unnecessary comment

* CMR-8665: Refactor function, add inflections dependency to transmit lib

* CMR-8665: inflections not needed

Inflections dependency is not needed in dev dependencies since it is included in the "dependencies"

* CMr-8665: refactoring function to make it more readable

Removing an extra space improving readability of a function

* CMR-8665: Passing spacing as a function

Pass the spacing as a function such that common app is unaware of of ingest or search

* CMR-8665: moving provider info to markdown cleaning up function definitions

* CMR-8665: Removing ingest references from common

* Adding a path to doc-routes function call for 3 arity
* moving provider information into the ingest.md file
* Creating options map to pass optional functions in ingest, search sites

* CMR-8665: Converting spacer to case function

Using a map is more simple in this case over a mathematical expression due to the rules for how many pounds are mapped to spaces being arbitrary but, essential to keep the table of contents formatting correct. Added tests for the case switch

* CMR-8655: Adding links for all concept types

* Create links for all concepts in the toc to match search's api.md
* Create headers for the endpoints to match search's api.md
* Fix granules to have their own toc section
* Add a few more headers for bulk updates to separate sections
* Minor fix to bulk updates status toc it is using a GET not a POST according to the section
* Encapsulates the functions in static to private