Bugfix CMR-10079: json access log (#2165)
The JSON Access log does not calculate duration correctly. Created a new Ring handler to capture the start time of a request. This handler is called as soon as possible in the Ring chain. Then the Log handler is to be called at the end of the chain and it uses the value to calculate duration.

While in the code, cleaned up the values reported in the log:
Removed values which are already in the headers
Added all headers
Added now-ms (current time as milliseconds)
renamed uri to url
added actual uri from request
changed remote address so that IP6 and IP4 localhost addresses report as "localhost"

What areas of the application does this impact?
Common, metadata-db, ingest, search, indexer, bootstrap, virtual-product, access-control
Not changed: mock-echo (port 2999)

A test was also changed so as to not issue warnings locally about using command function names (parse-double).